backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
langchain-core
langchain-ollama
opencv-python
eyepop
youtube-transcript-api
google-api-python-client
sqlalchemy
aiosqlite
cryptography
passlib[bcrypt]
python-jose[cryptography]
httpx
apscheduler
google-auth
requests
